Add files via upload
Tableau topic: "How to Build a Gantt Chart"

We generally use Gantt charts to show the duration of events or activities.

In a Gantt chart, each separate mark (usually a bar) shows a duration. 
One example use case of a Gantt chart is to display the average time to ship a range of products.


The basic building blocks for a Gantt Chart are as follows:
Mark Type: Automatic or Gantt Bar
Columns shelf: Date or Date & Time field (continuous measure)
Rows shelf: Dimension(s)
Size: Continuous measure
Color : Dimension (if required)

The distinguishing characteristic of Gantt Bars is that the length of every mark is proportional to the measure placed on Size on the Marks card.

The attached document provides the exact steps to create a Gantt Chart in Tableau without using the "Show Me" option.